+++ src/modules/mvuElement.css
@@ -1,5 +1,5 @@
 /* ==========================================================================
-   default style for all BaElements
+   Default style for all MvuElements
    ========================================================================== */
 *,
 ::before,
