+++ test/modules/footer/components/mapInfo/MapInfo.test.js
@@ -19,7 +19,7 @@ describe('MapInfo', () => {
 		});
 
 		it('contains child components with user selectable content', async () => {
-			// HINT: the existence of the behavior (user select text) is driven by css-classes specified in main.css and baElement.css.
+			// HINT: the existence of the behavior (user select text) is driven by css-classes specified in main.css and mvuElement.css.
 			// All elements are not selectable by default, but can be activated with the 'selectable' class.
 			const cssClass = 'selectable';
 			const element = await setup();
