+++ test/modules/geoResourceInfo/components/GeoResourceInfoPanel.test.js
@@ -57,7 +57,7 @@ describe('GeoResourceInfoPanel', () => {
 		});
 
 		it('should show a geoResourceInfo with selectable content', async () => {
-			// HINT: the existence of the behavior (user select text) is driven by css-classes specified in main.css and baElement.css.
+			// HINT: the existence of the behavior (user select text) is driven by css-classes specified in main.css and mvuElement.css.
 			// All elements are not selectable by default, but can be activated with the 'selectable' class.
 			const cssClass = 'selectable';
 			const geoResourceInfo = new GeoResourceInfoResult('<b>content</b>');
