+++ test/modules/layerManager/components/LayerManager.test.js
@@ -154,7 +154,7 @@ describe('LayerManager', () => {
 		});
 
 		it('renders draggable elements', () => {
-			// HINT: the existence of the behavior (user can drag an element) is additionally driven by css-classes specified in main.css and baElement.css.
+			// HINT: the existence of the behavior (user can drag an element) is additionally driven by css-classes specified in main.css and mvuElement.css.
 			// All elements are not draggable by default, but can be activated with the 'draggable' class.
 			const draggableClass = 'draggable';
