+++ test/modules/map/components/contextMenu/MapContextMenuContent.test.js
@@ -121,7 +121,7 @@ describe('OlMapContextMenuContent', () => {
 		});
 
 		it('renders selectable content', async () => {
-			// HINT: the existence of the behavior (user select text) is driven by css-classes specified in main.css and baElement.css.
+			// HINT: the existence of the behavior (user select text) is driven by css-classes specified in main.css and mvuElement.css.
 			// All elements are not selectable by default, but can be activated with the 'selectable' class.
 			const cssClass = 'selectable';
 			const coordinateMock = [1000, 2000];
