+++ test/modules/public/components/PublicComponent.test.js
@@ -63,7 +63,7 @@ describe('PublicComponent', () => {
 			const element = await setup();
 
 			//element.shadowRoot.styleSheets[0] --> main.css
-			//element.shadowRoot.styleSheets[1] --> baElement.css
+			//element.shadowRoot.styleSheets[1] --> mvuElement.css
 			//element.shadowRoot.styleSheets[2] --> publicComponent.css
 			expect(element._root.styleSheets.length).toBe(3);
 			expect(element._root.styleSheets[2].cssRules.item(0).cssText).toContain('contain: layout;');
