+++ test/modules/routing/components/Waypoints.test.js
@@ -114,7 +114,7 @@ describe('Waypoints', () => {
 		});
 
 		it('renders draggable elements', async () => {
-			// HINT: the existence of the behavior (user can drag an element) is additionally driven by css-classes specified in main.css and baElement.css.
+			// HINT: the existence of the behavior (user can drag an element) is additionally driven by css-classes specified in main.css and mvuElement.css.
 			// All elements are not draggable by default, but can be activated with the 'draggable' class.
 			const draggableClass = 'draggable';
 			const element = await setup(defaultRoutingState);
