+++ test/modules/toolbox/components/measureToolContent/MeasureToolContent.test.js
@@ -210,7 +210,7 @@ describe('MeasureToolContent', () => {
 		});
 
 		it('shows selectable measurement values', async () => {
-			// HINT: the existence of the behavior (user select text) is driven by css-classes specified in main.css and baElement.css.
+			// HINT: the existence of the behavior (user select text) is driven by css-classes specified in main.css and mvuElement.css.
 			// All elements are not selectable by default, but can be activated with the 'selectable' class.
 			const cssClass = 'selectable';
 			const state = {
